backend/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29.0
alembic>=1.13.0

# Haystack RAG
haystack-ai>=2.8.0
qdrant-haystack>=6.0.0

# LLM / Embeddings (OpenAI-compatible client for Ollama/vLLM)
openai>=1.0.0

# Document Processing
docling>=2.0.0

# Chunking
tiktoken>=0.7.0

# Task Queue (placeholder)
celery[redis]>=5.4.0
redis>=5.0.0

# Logging
structlog>=24.0.0

# HTTP Client
httpx>=0.27.0

# JWT validation
python-jose[cryptography]>=3.3.0

# Dev / Test
pytest>=8.0
pytest-asyncio>=0.23.0
ruff>=0.8.0
